requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
pywebview>=4.0.0
fastmcp>=0.1.0
scikit-learn>=1.0.0
watchdog>=4.0.0
